migrating to xmlschema
