+++ l10n_it_fatturapa/__manifest__.py
@@ -44,6 +44,7 @@
     'external_dependencies': {
         'python': [
             'pyxb',  # pyxb 1.2.6
+            'xmlschema',
             'asn1crypto'
         ],
     }
